6-size.c

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(a));
-printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(b));
-printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(c));
-printf("Size of a long long: %lu byte(s)\n", (unsigned long)sizeof(d));
-printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(f));
+printf("Size of a char: %d byte(s)\n", (int)sizeof(a));
+printf("Size of an int: %d byte(s)\n", (int)sizeof(b));
+printf("Size of a long int: %d byte(s)\n", (int)sizeof(c));
+printf("Size of a long long: %d byte(s)\n", (int)sizeof(d));
+printf("Size of a float: %d byte(s)\n", (int)sizeof(f));
 return (0);
 }
